backend/requirements.txt
# ============================================================
# Trading Bot - Requirements (OPTIMIZED VERSION v2.0)
# Todas as dependências atualizadas para Outubro 2025
# ============================================================

# API Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==13.1

# Binance API
python-binance==1.0.21

# Database & ORM
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
redis==5.2.0

# Data Analysis & Technical Indicators
pandas==2.2.3
numpy==2.1.3
ta==0.11.0

# Configuration & Environment
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1

# Telegram Integration
python-telegram-bot==21.8

# HTTP & Async
aiohttp==3.11.7
httpx==0.27.2
aiofiles==24.1.0

# Retry & Resilience
tenacity==9.0.0

# Testing (Development)
pytest==8.3.3
pytest-asyncio==0.24.0

# Utilities
python-dateutil==2.9.0.post0
